Add basic AST and parser  - Add delcaration, types and top to AST.  - Add parser.cpp for basic parsing.  - Add TokenCategory and mapping between it and tokenType  - Add logging util

+++ include/Lexer/Lexer.h
@@ -26,7 +26,8 @@ struct Token {
     if (other.type != type) {
       std::cout << std::format(
           "Token equality: type '{}' is not equal to expected type '{}'\n",
-          std::to_string(other.type), std::to_string(type));
+          std::to_string(static_cast<int>(other.type)),
+          std::to_string(static_cast<int>(type)));
       success = false;
     }
     return success;
